add -d option
spawning pmenu with xclickroot has the overhead of the process execution
and the initialization of Xlib, Imlib2 and the menu itself.  Thinkin on
that I created the -d option.  So instead of spawning pmenu when right
clicking and wait for it to initialize, now I only send a SIGUSR1 to
pmenu when right clicking.  Much faster.