backwards compatibility
I replaced "selbackground" with "activeBackground" on a previous commit
for compatibility with other X11 application ("activeBackground" is how
this resource is called on Athena and Motif widgets, "selbackground" is
not used anywhere else).

But since pmenu have used the name "selbackground" before, it should
keep recognizing it names for not breaking the configuration of anyone
out there depending on it.